Fix LogDetails structure to call API from inside component
