+++ frontend/src/pages/SocialFollowingPage.jsx
@@ -1,36 +1,17 @@
 import { useState } from "react"
 
-import { useInfiniteQuery, useMutation, useQuery } from "@tanstack/react-query"
+import { useQuery } from "@tanstack/react-query"
 import { styled } from "styled-components"
 
 import CommonCalendar from "@components/common/CommonCalendar"
-import { SkeletonProjectPage } from "@components/project/skeletons/SkeletonProjectPage"
 import SocialPageTitle from "@components/social/SocialPageTitle"
 import LogDetails from "@components/social/logDetails/LogDetails"
 import LogsPreview from "@components/social/logsPreview/LogsPreview"
 
 import { getCurrentUsername } from "@api/client"
-import {
-    getDailyLogDrawers,
-    getDailyLogsPreview,
-    getQuote,
-    postQuote,
-} from "@api/social.api"
+import { getDailyLogsPreview } from "@api/social.api"
 
 import { ifMobile } from "@utils/useScreenType"
-
-import queryClient from "@queries/queryClient"
-
-import { ImpressionArea } from "@toss/impression-area"
-import { toast } from "react-toastify"
-
-const getCursorFromURL = (url) => {
-    if (!url) return null
-
-    const u = new URL(url)
-    const cursor = u.searchParams.get("cursor")
-    return cursor
-}
 
 const SocialFollowingPage = () => {
     const initial_date = new Date()
@@ -48,46 +29,6 @@ const SocialFollowingPage = () => {
         queryFn: () => getDailyLogsPreview(me, selectedDate),
         enabled: !!selectedDate && !!me,
     })
-
-    const { data: quote, isPending: isQuotePending } = useQuery({
-        queryKey: ["quote", targetUser, selectedDate],
-        queryFn: () => getQuote(targetUser, selectedDate),
-        enabled: !!selectedDate && !!me,
-    })
-
-    const QuoteMutation = useMutation({
-        mutationFn: ({ day, content }) => {
-            return postQuote(day, content)
-        },
-        onSuccess: () => {
-            queryClient.invalidateQueries({
-                queryKey: ["quote", me, selectedDate],
-            })
-        },
-        onError: (e) => {
-            toast.error(e)
-        },
-    })
-
-    const {
-        data: drawerPage,
-        fetchNextPage: fetchNextDrawerPage,
-        isPending: isDrawerPending,
-        refetch: refetchDrawer,
-    } = useInfiniteQuery({
-        queryKey: ["daily", "log", "details", "drawer", targetUser],
-        queryFn: (page) => getDailyLogDrawers(targetUser, page.pageParam),
-        initialPageParam: "",
-        getNextPageParam: (lastPage) => getCursorFromURL(lastPage.next),
-    })
-
-    const hasNextPage =
-        drawerPage?.pages[drawerPage?.pages?.length - 1].next !== null
-    const isLogDetailsEmpty = drawerPage?.pages[0]?.results?.length === 0
-
-    const saveQuote = (content) => {
-        QuoteMutation.mutate({ day: selectedDate, content })
-    }
 
     return (
         <>
@@ -114,28 +55,10 @@ const SocialFollowingPage = () => {
 
                 {/* TODO: 날짜가 선택되지 않았을 때 */}
                 <StickyContainer>
-                    {isQuotePending || isDrawerPending ? (
-                        <SkeletonProjectPage />
-                    ) : (
-                        drawerPage && (
-                            <LogDetails
-                                user={quote?.user}
-                                quote={quote}
-                                saveQuote={saveQuote}
-                                selectedDate={selectedDate}
-                                logDetails={drawerPage}
-                                isFollowingPage
-                            />
-                        )
-                    )}
-                    <ImpressionArea
-                        onImpressionStart={() => fetchNextDrawerPage()}
-                        timeThreshold={200}>
-                        {hasNextPage && "next"}
-                        {!hasNextPage && !isLogDetailsEmpty && "no_more"}
-                    </ImpressionArea>
-
-                    {isLogDetailsEmpty && "empty"}
+                    <LogDetails
+                        username={targetUser}
+                        selectedDate={selectedDate}
+                    />
                 </StickyContainer>
             </Wrapper>
         </>
